Neural Networks Animal Classifer
This was my final project for a course in Neural Networks. I completed
the project in Fall 2011. The program implements a learning vector
quantization algorithm in order to classify what animal group an animal
belongs to based on their features. It also allows the user to choose
how they would like to initialize their weights for training. The
different methods for initialization are picking the first available
class of animal, random valued weights, zero weights, and using the
Kohonen SOM algorithm. I did research on which method provided the best
results. I found the best method of intilializing weights was a small
tweak of the Kohonen SOM algorithm. The download folder contains the
program, a file with animal data, and a program that can convert the
animal data into training and testing files.